Moved Amplitude module to `Amplitude/`
